Added the remaining classes and finished Flight, DirectFlight and IndirectFlight.

+++ src/Airline.java
@@ -1,13 +1,13 @@
 public class Airline {
     private String name;
-    private double fee;
+    private float fee;
 
-    public Airline(String name, double fee) {
+    public Airline(String name, float fee) {
         this.name = name;
         this.fee = fee;
     }
 
     public String getName() {return name;}
 
-    public double getFee() {return fee;}
+    public float getFee() {return fee;}
 }
